refactor popup requests
